Create TripController and include in main.js
